requirements.txt
browser-use
litellm>=1.10.1
openai>=1.3.0
python-dotenv>=1.0.0
protobuf~=4.24.3
httpx>=0.25.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
aiohttp>=3.8.5
asyncio>=3.4.3
typing-extensions>=4.7.1
pydantic>=2.4.2